requirements.txt
flask
psycopg2
gunicorn
dnspython
PyYAML